refactor: reorganize package structure for game interfaces
- Moved Playable, Replayable, and RoundBased interfaces to the 'interfaces' package for better organization and clarity.
- Updated package declarations in the affected files to reflect the new structure.
- Enhanced code maintainability by clearly defining the purpose of each interface within its dedicated package.
