+++ src/code/ca/bcit/comp2522/gameproject/interfaces/Playable.java
@@ -1,4 +1,4 @@
-package ca.bcit.comp2522.gameproject;
+package ca.bcit.comp2522.gameproject.interfaces;
 
 /**
  * Represents an entity or action that can be "played" or executed.
